Add files via upload
Assignment 1: Exploring Word Vectors (Coding Part)

In addition to a provided one, I used my own text in order to check my functions. I made word embeddings out of the 1st chapter of the "Doll", a book of a Polish author Boleslaw Prus.
It was only my attempt, therefore I do not claim that this solution is entirely correct or an only one.